+++ src/components/Search/Search.test.tsx
@@ -1,22 +1,18 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
 
-import { SearchInput } from './SearchInput'
+import { Search } from './Search'
 
-describe('SearchInput component', () => {
+describe('Search component', () => {
   it('renders without errors', () => {
     const mockSubmit = jest.fn()
-    const { queryByTestId } = render(
-      <SearchInput onSubmit={mockSubmit}></SearchInput>
-    )
+    const { queryByTestId } = render(<Search onSubmit={mockSubmit}></Search>)
     expect(queryByTestId('textInput')).toBeInTheDocument()
   })
 
   it('implements an onSubmit handler', () => {
     const mockSubmit = jest.fn()
-    const { getByTestId } = render(
-      <SearchInput onSubmit={mockSubmit}></SearchInput>
-    )
+    const { getByTestId } = render(<Search onSubmit={mockSubmit}></Search>)
 
     fireEvent.submit(getByTestId('textInput'))
     expect(mockSubmit).toHaveBeenCalledTimes(1)