+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.2.1] - 2020-05-01
 ### Develop -- Unrealeased
 
-- Added searchInput component
+- Added search component
 
 ## [1.2.0] - 2020-04-30